cleanup client support
