--- test/fixtures/renderContent/common/all.ejs
@@ -1,1 +1,0 @@
-common
