--- test/fixtures/renderContent/common/common.ejs
@@ -1,1 +1,0 @@
-common
