--- test/fixtures/renderContent/include.ejs
@@ -1,1 +1,0 @@
-<%- include('/all.ejs') %>
