--- test/fixtures/renderContent/include_common.ejs
@@ -1,1 +1,0 @@
-<%- include('/common.ejs') %>
