--- test/fixtures/renderContent/specific/all.ejs
@@ -1,1 +1,0 @@
-specific
